insert csv data into countriesStatistics.txt

+++ src/main/java/secondSet/countriesStatistics/Country.java
@@ -5,5 +5,5 @@ public class Country {
     String name;
     String continent;
     double area;
-    long population;
+    double population;
 }
